+++ README.md
@@ -1,9 +1,9 @@
 # tbdex-kt
 This repo contains 3 jvm packages:
 
-* [`/protocol`](./protocol/) - create, parse, verify, and validate the tbdex messages and resources defined in the [protocol draft specification](https://github.com/TBD54566975/tbdex-protocol/blob/main/README.md)
-* [`/http-client`](./http-client) - An HTTP client that can be used to send tbdex messages to PFIs
-* [`/http-server`](./http-server) - A configurable implementation of the [tbdex http api draft specification](https://github.com/TBD54566975/tbdex-protocol/blob/main/rest-api/README.md)
+* [`/protocol`](./protocol/) - create, parse, verify, and validate the tbdex messages and resources defined in the [protocol draft specification](https://github.com/TBD54566975/tbdex/blob/main/README.md)
+* [`/httpclient`](./httpclient) - An HTTP client that can be used to send tbdex messages to PFIs
+* [`/httpserver`](./httpserver) - A configurable implementation of the [tbdex http api draft specification](https://github.com/TBD54566975/tbdex/blob/main/rest-api/README.md)
 
 ## tbDEX Schemas
 > **Warning**
