+++ protocol/build.gradle.kts
@@ -20,11 +20,12 @@ dependencies {
   implementation("com.fasterxml.jackson.module:jackson-module-kotlin:2.14.2")
   implementation("com.fasterxml.jackson.datatype:jackson-datatype-jsr310:2.9.8")
   implementation("com.fasterxml.jackson.datatype:jackson-datatype-json-org:2.11.0")
-  implementation("com.github.TBD54566975.web5-sdk-kotlin:common:PR63-SNAPSHOT")
-  implementation("com.github.TBD54566975.web5-sdk-kotlin:credentials:PR63-SNAPSHOT")
-  implementation("com.github.TBD54566975.web5-sdk-kotlin:crypto:PR63-SNAPSHOT")
-  implementation("com.github.TBD54566975.web5-sdk-kotlin:dids:PR63-SNAPSHOT")
+  implementation("com.github.TBD54566975.web5-sdk-kotlin:common:main-SNAPSHOT")
+  implementation("com.github.TBD54566975.web5-sdk-kotlin:credentials:main-SNAPSHOT")
+  implementation("com.github.TBD54566975.web5-sdk-kotlin:crypto:main-SNAPSHOT")
+  implementation("com.github.TBD54566975.web5-sdk-kotlin:dids:main-SNAPSHOT")
   implementation("com.github.erosb:everit-json-schema:1.14.2")
+  implementation("decentralized-identity:did-common-java:1.9.0")
   testImplementation(kotlin("test"))
   testImplementation("com.willowtreeapps.assertk:assertk:0.27.0")
 }
