--- protocol/src/test/kotlin/models/OfferingTest.kt
@@ -1,56 +1,0 @@
-package protocol.models
-
-import assertk.assertAll
-import assertk.assertThat
-import assertk.assertions.isEqualTo
-import models.CurrencyDetails
-import models.Offering
-import models.OfferingData
-import models.Resource
-import org.junit.jupiter.api.assertDoesNotThrow
-import protocol.TestData
-import kotlin.test.Test
-import kotlin.test.assertContains
-import kotlin.test.assertIs
-
-class OfferingTest {
-  @Test
-  fun `can create a new offering`() {
-    val offering = Offering.create(
-      from = TestData.PFI,
-      OfferingData(
-        description = "my fake offering",
-        payoutUnitsPerPayinUnit = 1,
-        payinCurrency = CurrencyDetails("AUD"),
-        payoutCurrency = CurrencyDetails("BTC"),
-        payinMethods = listOf(),
-        payoutMethods = listOf(),
-        requiredClaims = listOf(TestData.getPresentationDefinition())
-      )
-    )
-
-    assertAll {
-      assertThat(offering.data.description).isEqualTo("my fake offering")
-      assertThat(offering.metadata.id.prefix).isEqualTo("offering")
-    }
-  }
-
-  @Test
-  fun `can parse offering from a json string`() {
-    val offering = TestData.getOffering()
-    offering.sign(TestData.PFI_DID)
-    val jsonResource = offering.toString()
-    val parsed = Resource.parse(jsonResource)
-
-    assertIs<Offering>(parsed)
-    assertThat(parsed.toString()).isEqualTo(jsonResource)
-  }
-  
-  @Test
-  fun `can validate an offering`() {
-    val offering = TestData.getOffering()
-    offering.sign("fakepk", "fakekid")
-
-    assertDoesNotThrow { Resource.validate(Json.stringify(offering)) }
-  }
-}
