--- protocol/src/test/kotlin/models/OrderStatusTest.kt
@@ -1,48 +1,0 @@
-package protocol.models
-
-import assertk.assertThat
-import assertk.assertions.isEqualTo
-import models.Message
-import models.MessageKind
-import models.OrderStatus
-import models.OrderStatusData
-import org.junit.jupiter.api.assertDoesNotThrow
-import protocol.TestData
-import protocol.TestData.PFI_DID
-import typeid.TypeID
-import kotlin.test.Test
-import kotlin.test.assertContains
-import kotlin.test.assertIs
-
-class OrderStatusTest {
-  @Test
-  fun `can create a new orderStatus`() {
-    val orderStatus = OrderStatus.create(
-      "pfi", "alice", TypeID(MessageKind.rfq.name), OrderStatusData("my status")
-    )
-
-    assertk.assertAll {
-      assertThat(orderStatus.metadata.id.prefix).isEqualTo("orderstatus")
-      assertThat(orderStatus.data.status).isEqualTo("my status")
-    }
-  }
-
-  @Test
-  fun `can parse an orderStatus from a json string`() {
-    val orderStatus = TestData.getOrderStatus()
-    orderStatus.sign(PFI_DID)
-    val jsonMessage = orderStatus.toString()
-    val parsedMessage = Message.parse(jsonMessage)
-
-    assertIs<OrderStatus>(parsedMessage)
-    assertThat(parsedMessage.toString()).isEqualTo(jsonMessage)
-  }
-
-  @Test
-  fun `can validate an orderStatus`() {
-    val orderStatus = TestData.getOrderStatus()
-    orderStatus.sign("fakepk", "fakekid")
-
-    assertDoesNotThrow { Message.parse(Json.stringify(orderStatus)) }
-  }
-}
