+++ settings.gradle.kts
@@ -1,2 +1,2 @@
 rootProject.name = "tbdex"
-include("http-client", "http-server", "protocol")
+include("httpclient", "httpserver", "protocol")
